+++ src/main/java/com/cclife/registration/domain/Family.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ws.cccm.application.model;
+package com.cclife.registration.domain;
 
 import java.io.Serializable;
 import java.util.Date;
@@ -62,6 +62,8 @@ public class Family implements Serializable {
     private String familyName;
     @Column(name = "HomeAddress", length = 60)
     private String homeAddress;
+    @Column(name = "HomeAddress2", length = 60)
+    private String homeAddress2;
     @Column(name = "HomeCity", length = 30)
     private String homeCity;
     @Column(name = "HomeState", length = 2)
@@ -250,6 +252,20 @@ public class Family implements Serializable {
     @Override
     public String toString() {
         return "ws.cccm.model.Family[ familyID=" + familyID + " ]";
+    }
+
+    /**
+     * @return the homeAddress2
+     */
+    public String getHomeAddress2() {
+        return homeAddress2;
+    }
+
+    /**
+     * @param homeAddress2 the homeAddress2 to set
+     */
+    public void setHomeAddress2(String homeAddress2) {
+        this.homeAddress2 = homeAddress2;
     }
     
 }
