+++ src/main/java/com/cclife/registration/domain/Person.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ws.cccm.application.model;
+package com.cclife.registration.domain;
 
 import java.io.Serializable;
 import java.util.Date;
